+++ schema/sqlite.sql
@@ -7,6 +7,7 @@ CREATE TABLE flysystem_path (
   size INTEGER DEFAULT NULL,
   is_compressed INTEGER DEFAULT 1,
   expiry TEXT DEFAULT NULL,
+  meta TEXT DEFAULT NULL,
   update_ts TEXT NOT NULL DEFAULT 0,
   PRIMARY KEY (path_id)
 );